add support on habana platform (#131)
* add habana

* add mask

* fix mask in outer_product_mean

* add dap

* add hmp

* merge training code

* add chunk for inference

* fix extra-msa stack for training

* support ddp in training

* fix inference bugs

* code refactoring for habana

* support hmp training

* enable all inference and train on Gaudi/Gaudi2 with optimized perf with latest base (#139)

* enable all inference and train on Gaudi/Gaudi2 with optimized perf

* refine code to adapt new base

* refine code to fix issues in code review

+++ fastfold/model/nn/embedders.py
@@ -360,7 +360,7 @@ class TemplateAngleEmbedder(nn.Module):
         Returns:
             x: [*, N_templ, N_res, C_out] embedding
         """
-        x = self.linear_1(x)
+        x = self.linear_1(x.to(dtype=self.linear_1.weight.dtype))
         x = self.relu(x)
         x = self.linear_2(x)
 
@@ -446,6 +446,6 @@ class ExtraMSAEmbedder(nn.Module):
         Returns:
             [*, N_extra_seq, N_res, C_out] embedding
         """
-        x = self.linear(x)
+        x = self.linear(x.to(dtype=self.linear.weight.dtype))
 
         return x
